weather/playwright-test/amazon-invoices-main/requirements.txt
greenlet==3.1.1
playwright==1.49.0
pyee==12.0.0
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
